Main change: the attribute editor now displays several types of attributes in a type-specific way, or in a way specified by the displayer type field of the attribute definition.
The displayer type 'hidden' means that an attribute is not to be shown
at all.

Window position and size attributes of all kinds are no longer shown,
nor is arm duration or the 'visible' flag.

The file selector code in NodeInfo has been moved to a new module,
selector.py, which does some clever things with current directory and
/dev/null, and is also used by the attribute editor for file names.

Module ChannelMap exports a new value, 'channeltypes', used by other
modules that need to display a menu of channel types (e.g. the 'new
channel' operation in ChannelView.py and the channel type field in the
channel attribute editor.