The MpegChannel is now officially gone.  For backward compatibility, a document that uses the MpegChannel will get the VideoChannel instead (which on most platforms can play MPEG videos).
There is now one VideoChannel.py which looks at sys.platform and
imports a platform-dependent version (called xxxVideoChannel where xxx
is SGI, MPEG, MAC, or NT).  These files can all be in the common
directory or in the platform dependent directory.
The platform-dependent VideoChannel should define a class VideoChannel
that implements the channel.

The change in ImageChannel is just a removal of superfluous space
characters.