Make image channels always transparent.  This means that if the width/height wasn't specified, the channel can extend to the edge of the parent channel.  This in turn means that we don't need to know the size of the image, as long as we know the size of the root-layout (so parsing the document is faster).
The consequence of this is that if an image has transparent parts, the
underlying channels will shine through, instead of the background of
the channel.  Also, if a widht/height is specified which is greater
than the image's width/height, you don't see the channel's background
color to the side/below the image but instead you see the underlying
channels.