This commit represents a MAJOR change, therefore it is committed in a branch (called "usercmd").
The intention is that the changes made here find their way into the
main branch.  Currently only the Motif version has been tested (and I
doubt that other versions work :-), so that is another reason to
commit this in a branch.

Now for the changes.

A new module "usercmd" was created (both for editor and player) that
contains a list of user-visible commands.  The commands are
represented by classes.  All user-interface modules (i.e. Player,
HierarchyView, ChannelView, and Channel) use this module to do two
things:
1. There is a list of commands with callbacks.  These are the commands
   that the module recognizes and can act upon.  The list of commands
   is given to the windowinterface which uses them to enable/disable
   certain commands (menu entries, buttons and such).  This part is
   (supposedly) platform-independent.  The list consists of instances
   of the classes defined in the usercmd module.  These instances are
   created with a "callback" instance variable which gives the
   callback function plus arguments to call.  The instances also have
   a "help" instance variable which is either None or a string with a
   one-line help message.  This message may be used (and on Motif is
   used) to display a small pop up window when the mouse is over the
   button or menu entry that the command refers to.
2. There is a desription of the desired user-interface.  In Motif this
   looks like a dictionary which describes the menubar, the toolbar,
   the shortcuts (accelarators), and what to do when the window
   manager close function is used on the window.  The implementation
   of this description is platform-dependent, but the fact there their
   is such a description is platform-independent.
   This description incorporates (some of) the classes from the
   usercmd module.  The windowinterface makes the link between the
   allowed commands (instances of these classes in list 1 above) and
   the user-interface elements, and calls the callbacks of list 1 when
   the interface element is selected.

The Channel with top-level windows ask the Player about the adornments
their window is supposed to get using the method get_adornments.  The
Player returns an opaque object (in Motif it's the dictionary
mentioned above), and this object is passed to the newwindow
function.

In the Motif editor, the only "adornment" is what to do when the
windowmanager close function is used.  In the player, this is used to
define a menubar and toolbar for the first toplevel window of each
document.  If no channels exist with a window, the players Player
creates it's own window with the same menubar and toolbar, but with
nothing more than that.

It is envisioned that the Mac uses this mechanism to create a single
menubar for the whole application.  The various windows can then
"subscribe" to this interface elements from this menubar using the
command list mentioned under 1 above.  The current Mac implementation
already does something similar to this.

It is also envisioned that the Windows implementation does something
between what the Mac does and what happens on Unix.

There are some changes to the windowinterface API.  The newwindow and
newcmwindow functions no longer have a "menubar" optional argument,
but instead have an "adornments" optional argument.  The Motif version
also uses "commandlist" and "resizable" optional arguments, but these
are only used in the platform-dependent part (I think :-).

Another change that was made:  There is now a MAGIC_PLAY command in
the player.  This is bound to the space bar.  If the document is
stopped, the document starts playing.  If the document is playing, the
space toggles between pausing and not pausing.

There is a new program mkimg.py in lib which creates Python modules
from image files.  This is a generalization of the mksplash.py program
which now calls mkimg to do the work.