Simplified cursor handling.
There is a new function in windowinterface setwaiting().  This
function is called to tell the windowinterface that it should change
the cursor into a watch cursor.

Callback functions call TopLevel.setwaiting() when they know that the
callback is going to take some time.  All this setwaiting call does is
call windowinterface.setwaiting().

It is the responsibility of the windowinterface to change the cursor
back to the current cursor when it returns to the main loop.

This means that there is no need for a call to TopLevel.setready(), so
all these calls have been removed.  Also, since the windowinterface
sets the cursor to a watch and back in all windows it knows about,
there is no need for TopLevel to recursively tell all views to change
the cursor, so all calls and definition of setwaiting and setready
have been removed (except for TopLevel.setwaiting, obviously).

There is still a windowinterface.setcursor function and setcursor
methods on all windows.  These are now not used anymore to set a watch
cursor, but they can still be used to set other cursors.

(The Motif implementation uses the setcursor method internally, but
even there it doesn't pass 'watch'.)

The Mac and Windows versions still need to implement the new cursor
policy and setwaiting function.