Implemented User Groups.
User groups are a way to easily group nodes in a logical way so that
it is trivial for the user of the document to select what he wants to
see and what not.  See Dick's paper in the proceedings of MM'98.

This is a first implementation and I am not yet completely happy with
the user interface for editing user groups (the UsergroupView*.py
files).  Most of the rest seems to be ok.