New way of interacting with the Structure View. The old version was tagged with "old_hierarchy_view".
Interior nodes are collapsed and expanded explicitly, and the canvas
size is adapted to fit everything.  There are no longer commands
available to zoom in or out, or to double the size of the canvas or
reset it to normal.

Expanding/collapsing nodes can be done through the menu (both menubar
and popup), or by clicking on a little triangle that also serves to
indicate the expand/collapse state of the node.