Version 1.0.1 Alpha: time for experimenting.
Added some attributes to hopefully make RealPix editing easier:
Subregion is split into x,y and w,h properties.
There is an "anchor" property which tells which corner of the
subregion is described by the x,y coordinates.
There is a "displayfull" boolean property which tells whether to use
the full region or to use the given x,y and w,h values.
When a new slide is created using drag-and-drop, the w,h is filled in
with the dimensions of the image, but displayfull is set to "on" so
that these values are not actually used.
Together this gives the user the dimension of the image which they can
use if they so desire, but the default behavior is unchanged.