Moved some of the RealPix specific code from HierarchyView to a new module.
Made all URLs in slides relative to the SMIL file instead of the
RealPix file.  This simplifies a lot of places by removing special
case code and complicates only two: reading and writing RP files.
This is also a prerequisite for having RP nodes without URL.

Now always create a SlideShow instance for RealPix nodes, instead of
doing it "on demand" (when the node is expanded, and at various other
places).  This required that there be a delayed editmgr.register()
since when the file is read in there is no editmgr, but we need to
register the SlideShow instance when it is created.  The delayed
register is implemented in the node context.

Fixed realconvert.convertimage so that it also works in a pure img
environment (Mac/Unix).

Fixed the return value of the OK callback of the Export command so
that it doesn't return a value after an error.  If a value is returned
that evaluates to true, the file dialog is not removed, which is not
what was wanted.