Experimental stuff. Make, a maybe temporal, cvs sketch of the current state of windows media sdk. The indent is to make clear beyond MS marketing smoke what is really available and what's is state.
Add automation support for windows media encoder tool.
Automation from python of COM entities exposing their interface through the generic IDispatch
is a broad and generic subject for which we are not concerned here.
Windows media encoder tool automation, is called by MS marketing an SDK but is certainly not yet.
Here we are just experimenting.