Complete rma pyd new implementation.
I think that it fulfills all the requirements I posed to myself:
The new implementation is more clean and can be extended more easily.
Errors and failures are catched and reported the python way.
Composition is done in python and thus the new rma is more a lib than an application (i.e playing is supported)
Dependency on pywintypes.dll of the previous win32 rma has been removed (this one depends only on python16.dll)