The semantics of the animation module force as to introduce two new entities: EffectiveAnimator and AnimateContext.
An EffectiveAnimator is responsible to combine properly
all animations of the same attribute and the base value
to give the final display value.
This is the entity that knows and keeps the attr display value
taking into account all animations and the dom value.
It implements animations composition semantics
	'additive' attribute + priorities
and is an entity at a higher level than animators (and thus channels).

AnimateContext is an EffectiveAnimator repository
We need a well-known repository so that we can find EffectiveAnimators
from Animators (channel) context.
It implements also operations that apply to all EffectiveAnimator objects
and is a document level entity.

Both EffectiveAnimator and AnimateContext are in skeleton form for now.
Their implementation seems straight and will follow.
The important thing is to realize that their introduction is needed (simplifies things both conceptually and implementation)
and to specify their responsibilities.

Also, first approach to implementating the 'accumulate' attribute.