Fix the first and most important source of instabilities of video on win9x. (the second part will follow)
Use directly surfaces bliting for video instead of using the gdi
(this was also my original implementation and has advantages even on NTx)
There must be a bug on win 9x implementation of the related infrastructure.
We use the same operation succesfully in many places but not with video.
Also be a little more carefull on the details than I do usually.