add a 'region' inherited type. It's basicly the same as 'channel' type, but instead, inherit from layoutchannel Currently, there is  some hack in MMChannel.__getitem__, ... to get the equivalent. In addition, it doesn't work for many complex cases. (I have checked) I absolutly need to introduce this new type in order to implement clearly the CSS2 behavior for at least the backgroundColor. Remind: in CSS2 and therefore in SMIL, for backgroundColor, we can have either:
- real color value
- transparent value
- inherited value

And at last, the default value is 'transparent' (not 'inherited').

The translation into internal GRiNS structure will be:

real bg color --> bgcolor = tuple, transparent = 0
transparent --> bgcolor = xxx, transparent = 1
inherited  --> bgcolor = None, transparent = None

Also, about backgroundColor, we'll able to keep the compatibility with SMIL1-G2 document with the the attribute 'showBackground'. For not hierarchical region (it's the case in SMIL1, it's the same behavior)

Note than transparent value is different as inherited value for some cases.