New file. Architectural change. Again.
This file will contain base classes used by the StructureView, TimelineView
and TemporalView. These will be objects for nodes and regions which provide
pop-up menus, editing functionality, menu bindings, command bindings, node
traversals and stuff like that. They will _not_ do any drawing code - this is
meant to be done by widgets. These will probably also be interfaces for the
edit manager.

In the M-V-C OO pattern, these are controllers. Widgets are the Views and
MMNode/MMChannel are the models. This is a very rough, hacked design, but
it will make things a bit easier.