Work for the last 4 days.
Replaced MMNode with EditableObjects. For the editor, EditableObjects do
lots of useful things like cut, paste, delete and so forth. For the player,
they are are an empty class that inherits from MMNode.

This works on the windows platform.

This change only affects the temporal view. This is an intermediate state -
I'm going to be digging my teeth into the other views ("because I can") soon.

To undo this change, use the tag "before-editable-objects".