management of channel renderers (leaf channels) from PlayerCore module
Now, all channel renderers are completely transparent for any editing operation. In order:

- it fixes a channel conflict problem you could have before after an editing operation
- it simplifies a lot some editing codes

the renderer channel allocation is the same as before ( in SMILTreeRead), but is recomputed each time you edit from PlayerCore.
The time needed to do this is about 0.2 s on my computer (so invisible)

Unfortunatly, there are still some minor hacks in PlayerCore to keep the player working. But it's very localized, and doen't affect the editor code. So, It shouldn't be a priority
I tried to remove these hacks, but it affects too many others part of code (sound channels, scheduler init, ...)