suppot for 'copy/cut/paste' regions
I tried to keep the same way as for mmnode hierarchical. But the current code is a bit confusing for me. And currently when we copy a node/sub-tree, the nodes copied into clipboard keep a link with the current context. It's a bit dangerous, and doesn't work currently in somes cases
So, I couldn't keep exactly the same way. I think, we'll have to discuss about these issues when Jack will come back ?