1) fix the source of the sound bug:
- call setlayout to force all channels to show state after that the renderers are made
before, all channels was correctly created, channel.play was called but interupt because the show method on the sound channels wasn't called

2) fix a new major editor bug introduced by the last fix

- you couldn't change the media type associated to a node anymore
- you couldn't change any media region anymore
...

all these problem was due to:

the channels wasn't recreate when there was an edit operation involving a region/channel changement

Of course, the current algorithm is not completly optimized. But the time requiered to re-make the renderer channels is quite short (max 0.5 s for BBC demo). And the fact to introduce a more complex algorithm and not re-create all renderer channels may requiered more times !, and may introduce some bugs difficult to find.