big speed improvement to generate the SMIL source: about two times faster: It becomes almost acceptable for BBC demo
more details:

currently, when a node is generated, all SMIL2 attributes are check to see if there is something to write.
we have today about 85 SMIL attributes for the body elements.
A easy way to optimize this was to call getcmifattr, (and other similar method ) only if the attribute was stored in x.attrdict in testing directly in the loop with the method has_key. Thus, it avoids many method called which are pretty long in Python.

Of course, there was still better solution but requiered too much work to gain only a little bit with our current framework.