don't call RecalcLayout when minimize: A easy way to fix the toolbar position problem when you minimize.
more details:

RecalcLayout allows to re-position all child windows (MDI child, tool bars, ...) when the main frame is resizing (see MDI doc for more details).
In our cases, the method was also called when the main frame is minimized. Thus, the tool bars was automaticly repositionned according to the new parent geometry (the new parent geometly should have been the icon size).
In addition, the toolbar behavior is changing position if the toolbars don't fit anymore inside the MainWindow. So, for the icon size, the toolbar positions couldn't fit anymore and was completly reseted.

There was maybe a better solution to fix this problem in changing the default method(s) implementation which are responsible to recalculate the new toolbar size. But it looks a lot more complex.