Basic stuff from WinInet SDK.
Enable a solution on windows for the following important bug reported by Dick:
"...there is no way to cancel the load: you need to kill GRiNS."

Of course we need asynchronous sockets or threads
for a proper solution.

With the help of this module we can know whether our user
is connected and if not initiate a connection.