Implemented optional pruning of branches of the tree that are unplayable with the current settings.  This is now a special case of the Save As command (but using a different menu entry and command), but could also be a special case of the export command.
Also fixed a problem when the disk becomes full while publishing:
before a message popped up warning about the problem, and the only
recourse you had was to click OK so that for the next file you got the
same message.  Now there is a cancel button as well.  Maybe we should
do some cleanup when the user cancelled...