MASSIVE CHECKIN!
Replaced anchorlist attribute with MMNode instances of type "anchor".

Reading and writing SMIL files works.
Writing XHTML+SMIL doesn't work yet.
Anchor attributes can be edited by selecting the anchor in the
Hyperlink window and clicking on Anchor editor.  Once media nodes can
be uncollapsed in the structure view, they will also be editable
through the structure view.
Most of the playback works, but there are still some details to be
ironed out.
Animating anchors has not been tested and presumably does not work.

Also still to be done: animate children of anchors (yes, that's
allowed in SMIL 2.0).

Also fixed: support for fragment attribute on HTML media (but this
only works when using the Webster HTML control -- the IE control only
seems to work when reading the HTML from a file).  Only tested on
Windows, but implemented on Mac and Unix as well.