animation editing: make things working for RP9:
- don't generate keyTimes attribute anymore which is not supported and won't be probably not supported in RP9 since keyTimes is not part of SMIL2 language profile:
use instead multiple animation nodes.
- update the timing information on animate node when the duration attribute on the media which is animated has changed. just to be able to demonstrate GRiNS. Of course it works just for a particular timing case.
This update happens only when the layout view is open, and the node which is animated is selected in order not to affect the entire editor stability

Note: a part of this node is purerly experimental, and another part should split/move and clean-up next week.