make working HTML+SMIL export when animate nodes are a sub-node of the media. IE5.5- IE6 didn't play correctly that kind of animation.
Also. It works for a limited number of cases. The main problem not resolved is:

the media size after applied the fit attribute has to be the same as the sub-region size. In short:
be carefull if you use fit different from 'fill' value, and animate it. IE/HTML fits by default any media into the entire region (fit="fill")

Note that this problem doesn't exist when the media is not animated (or a parent region), because the export code determinate staticly the right region size according to the fit attribute.