Start implementing an infrastructure for the animation editing stuff by using a part of codes what we have currently. Important: Keep working the current implementation until the new will be finished. I added a temporare flag in layout view for that.
This new stuff is especially important for the following raisons:

- To avoid the timing problems we have now
- To improve the reliability
- Be able to add easily new features. For example, not to merge anymore the animate node and its target node, ...

Details:

Move a part of AnimationData into a new class for interpolation facility, and eventually some other utilities for later.
Another part will be soon moved and cleaned into the modules which save/publish the document.
The part that Sjoerd wrote will be used for playing.
At last the most experimental part we have (SMIL animation import) currently won't be used for now. Eventually much later to be able to export full SMIL2 animation after a big clean-up.