Made parameters of SMILWriter (SMILTreeWrite.WriteFile and friends) more orthogonal: - grinsExt governs GRiNS extensions and GRiNS namespace; - rpExt governs Real extensions and Real namespace; - qtExt governs QuickTime extensions and quickTime namespace; - copyFiles says whether or not to copy files; - convertfiles says whether or not to convert copied files to   RealMedia; - convertURLs says whether of not to convert URL to be relative to   filename. Save to .grins file has grinsExt on and rpExt and qtExt on depending on product (whether or not the feature is in the feature_set). Save to .smil file has grinsExt off. Publish has at most one of qtExt and rpExt on (depending on what you publish to) and has copyFiles on.
Removed the term G2 from the user interface and replaced with
RealPlayer or RealMedia.

Added a feature EXPORT_REAL so that it is easy to enable/disable.

Export/Publish to G2 is now Export/Publish to RealPlayer, and uses
SMIL 1.0 (RealPlayer G2) or SMIL 2.0 (RealPlayer 9) syntax depending
on whether the document is a SMIL 1.0 or SMIL 2.0 document.

Cleaned up command list stuff in TopLevel.

Cleaned up some stuff in SMILTreeWrite.