fix the random focus behavior bug and all problems related:
To solve the problem, we have now two invisible control (instead to one before).
The first one allows to manage the pane focus since the pane can't be a part of the general focus management from Windows. In fact when the pane is Hilighted, it's not the pane itself which has the focus but this invisible control. Also, when the pane get automaticly the focus by Windows, it is redirected to this control.
The second invisible control allow to manage a part of the tree widget focus. When this control has the focus, it is transferted to the tree widget