New experimental project: GRiNS-Boost:Bridge
GRiNS-Boost:Bridge project task is to separate completely the UI of the player and factor out GRiNS player engine as is. Especially under windows, it?s important to break application's strong coupling with MFC (win32ui, pythonwin). In a sense this project can be seen as an extension to our previous WGBH java project and also to the more than a year long os-windowless-regions redesign efforts.  This will help us to add value to our player by extending its applicability domains (i.e. to areas not easily reachable by Real). Will help us also make porting to CE and other devices a much simpler task.

I think this task is feasible and I hope that it will be proven useful. And anyway, it is already complete at a level we can see the benefits and drawbacks.